feat: (BREAKING) Rename overwrite and add overwrites variables
This is because the way the `overwrite` argument on the resources works
was badly explained and interpreted. The overwrite parameter on the ssm
parameter resource was meant to be used only during initial creation of
the terraform resource. The attirbute is set that if the parameter
already exists when terraform tries to create it, it shouldn't be
overwritten. It's now also clearly stipulated in the documentation that
if you want to manage the update mechanism of the value within terraform
a life cycle rule should be used.

Therefore, I'm renaming the overwrite variable of the module to
ignore_changes_on_value and am adding the overwrites variable as a list
of bolleans that can be set individually on each ssm parameter resource.

This will make the module less confusing for users and will use the
resource as the provider intended the argument to be used.